+++ app/Livewire/Events/EventEdit.php
@@ -84,7 +84,7 @@ class EventEdit extends Component
         NotificationMessage::addNotificationMessage(
             new Item(__('The event has been successfully updated.'), ItemType::SUCCESS));
 
-        return redirect(route('event.index'));
+        return redirect(route('events.index'));
     }
 
     /**